+++ src/helpers/variableType.js
@@ -11,7 +11,7 @@ export function isString(variable) {
  * @returns {boolean}
  */
 export function isNumber(variable) {
-	return (typeof variable === 'number') || (variable instanceof Number);
+	return ((typeof variable === 'number') || (variable instanceof Number)) && !Number.isNaN(variable);
 }
 
 /**